(chore) add local development configuration, including dev docker and logs
