+++ README.md
@@ -1,4 +1,6 @@
-# A naive RFC-8984 Rust implementation
+# BackendSouls Calendar
+
+A naive Rust implementation of the RFC-8984!
 
 ## Dependencies
 
@@ -32,4 +34,4 @@
 
 ## Contact
 
-Signal: [RFC-8984 Group](https://signal.group/#CjQKILpMo2tZkR1k6iKem8WQd9l2YN2EEMQs11RzYNDe9LxcEhDOc5uyJEom2C2C-quR7Uq3)
+Signal: [BackendSouls Calendar Group](https://signal.group/#CjQKILpMo2tZkR1k6iKem8WQd9l2YN2EEMQs11RzYNDe9LxcEhDOc5uyJEom2C2C-quR7Uq3)
