+++ modules/calendar-api/src/lib.rs
@@ -1,5 +1,6 @@
 pub mod app;
 pub mod configuration;
+pub mod environment;
 pub mod event;
 pub mod group;
 pub mod task;
