+++ modules/calendar-domain/Cargo.toml
@@ -4,3 +4,6 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
+derive_builder = { workspace = true }
+anyhow = { workspace = true }
+getset = { workspace = true }
